buildsystem tweaks: make the python interpreter configurable
Up till now the buildsystem used the first "python" in PATH. That makes it
difficult to use a specific interpreter if one has several around.
Now one can set the "PYTHON" variable in build.conf i.e. to
"/opt/local/bin/python2.6" to build agains python2.6.
If this variable is not set, the old behaviour stays in place.

Note: This is only for building on unix. On windows, the "PYTHON"-variable is
not used, but the interpreter set in the windows registry to execute
".py"-files.

git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1404 5a6a7de5-fb12-0410-b871-c33778c25c60