Python interface: added Table.setRowData() and setColData()
The first one takes a row index (int) and a Python tuple with the new values
for each column.
The second takes a column indes (int or str) and a Python list of values for
this column, one per row. An optional third argument can be given as row
offset. Example: `table("Table1").setColData(2, [i*i for i in range(40)], -1)`
If the list argument ends before the end of the table is reached, a
`StopIteration` is raised. In conjunction with the row offset, this allows to
fill a column chunk wise.
For now, only numeric and text columns are supported.

git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1428 5a6a7de5-fb12-0410-b871-c33778c25c60