Python interface: added Table.addRow() and Table.appendRowData()
The first method will simply add a new empty row at the end of the table and
return the index of the new row.
The second one also creates a new row at the end of the table, but it will
call setRowData to fill the new row. The data is passed as Python tuple with
one element per column.

git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1429 5a6a7de5-fb12-0410-b871-c33778c25c60