Python interface: added Table.cellData and Table.setCellData
These methods can work on abitary python types and cast from/to the type of the
corespondent column. The implementation shares the code with rowData/colData
and setRowData/setColData.
For now only the numeric and text types are supported. Support for ither types
can be added to cellDataHelper and setCellDataHelper in the %TypeHeaderCode

git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1450 5a6a7de5-fb12-0410-b871-c33778c25c60